Adding NoneGroup spec group type
Indicates the uploader should not attempt to collect anything from this
spec.  This is to provide separate definitions for different formats of
the same command.  This generally happens when one format is collected
by SoSReport, but Insights collects a different version (usually to get
more or more easily parseable information).

lvs and pvs are the first commands to use this spec group.