Add optional keyword to @reducer
This gives us the ability for plugins to indicate which mappers they
use but should still execute if it's not present.  This enables us to
better determine which files to collect on client machines based on the
requirements of installed plugins.

Previously file collection was determined based upon which mappers had
been imported after importing all installed plugins.  This doesn't work
if more than one mapper are defined in a particular module and one
mapper is used by a plugin whereas the other one is not.

The syntax is similar to @reducer.requires:

from falafel.mappers.optional import OptionalMapper
@reducer(optional=[OptionalMapper])
def report(local, shared):
    pass