Changing to using a state machine for parsing.
Instead of using a global per-file variable to store the position in each list
of lines and multiple for loops to parse the file, the parsing code is in
sub-functions and we use a state machine in a single loop to use the right
parser for each section we're reading.  Hopefully this is a bit nicer than the
previous method.