Add -show-plugin-modules option to insights-cli
The --show-plugin-modules option tells insights-cli to list
all the plugins modules for Insights. A python package needs to
have 'insights-rules' tag in Keywords metadata to identify it as
a Insights plugin package.

A keyword-argument `keywords='insights-rules'` needs
to be added in setup() call in setup.py file of the rules
repository.