Take a 'multikeys' parameter which allows multiple values to be stored.
Normally, if a key occurs more than once in the same section the later value
will replace the earlier value.  However, if this key is listed in the
'multikeys' list this will cause these values to be appended to a list.  If
multikeys is set to '*', all keys will do this.  The key names to allow as
multiple value keys can be given as anything to which the 'in' operator
applies - a list, dictionary or string.