Adding missing requirements to evaluator response
This required we add a new response type called "skip".  The data
structure attached to the evaluator response is a list of dicts that
looks like this:

{
    "type": "skip",
    "reason": "MISSING_REQUIREMENTS",
    "details": (missing_all, missing_any)
}

where "missing_all" is a list of reducer class objects, and
"missing_any" is a list of lists, where each list is a set of mappers
that represent one single requirement.

InsightsEvaluators stringify the missing requirements since class
objects aren't JSON serializable.