Add basic caching to content manager
Will load all content from filesystem once and keep it in an internal
dict.  There is a new method to refresh the content.

Also adding watchdog as an optional dependency/feature.  This will watch
the filesystem while the content server is running and automatically
reload the content from the filesystem when any changes are detected.