A common 'SysconfigOptions' class for parsing /etc/sysconfig/ files
Several mappers implement a parser for /etc/sysconfig/ files, but they
often get several things wrong in comparison to bash, which is the usual
interpreter of these files:

* They don't remove double quotes and single quotes around strings.
* They don't ignore comments starting with a hash.
* A blank value is acceptable (e.g. 'OPTIONS=' sets the value to '')

So in order to reduce code duplication, make the parsing more accurately
reflect what bash will do, and to provide better test coverage for the
general case, I have written a 'SysconfigOptions' class which these and
other mappers can subclass.  This format also matches the systemd
'EnvironmentFile' definition.  This uses the 'shlex' module to parse
quotes and comments in the same way that shells like bash do.

All options read from the configuration file are available in the 'data'
attribute of the instance.  It is suggested that for commonly checked
variables, subclasses define a property that fetches that variable with
a suitable default in case the variable is not defined.

Also made the new KexecCrashSize mapper a little simpler.