Changing the CobblerModulesConf to an IniConfigFile mapper
This mapper used to reimplement '.ini' file parsing, badly.  It now uses
the standard `IniConfigFile` mapper class.

Tests have been updated.

No rule in the CEEPH repository uses this mapper.  The only use of it is in
the diag-insights-rules repository.  Changing that to use the IniConfigFile
format will be done shortly.

This commit also adds documentation including sample data and example code.