DockerList mappers with sample data and usage code
The original mappers were functions that were hard-coded to their specific
formats, didn't cope with other fields being blank, and only returned lists.
To be more versatile, the functionality was rewritten as a class that
looked at the header of the output and used the field descriptions and
column sizes from that to construct the row data lists that are stored in
the `rows` attribute.

If a subclass sets the 'key_field' to a string that is in the header, the
data will also be stored in a `data` attribute that is a dictionary, where
each row is stored keyed to the value of the key field given.

The two mappers are now provided in the `DockerListImages` and
`DockerListContainers` classes.

The original mapper functions are still provided, and retrieve the data
using the DockerList classes and return the `rows` list as that's what
the reducers and tests expect.