Adding VDSMConfIni mapper and refactoring VDSMConf to use that
Yet another mapper parsing INI files by hand into a dict of dicts,
so we now use IniConfigFile for the actual parsing and for its many
other affordances.  However, since VDSMConf just uses a single dict
of dicts in the 'data' property, which is incompatible with the
IniConfigFile structure, we construct the dict of dicts using a
dictionary comprehension.

Documentation and tests have also been added.  The CEEPH rules have
also been checked and all pass.