Correctly parse file names ending with a colon
The FileListing directory parser looked for new directories by finding
lines that ended with ':'.  However, this is a valid, if unusual, file
name ending.  So we now detect that the directory name starts with '/'
as well, since the file type attribute cannot contain a '/'.  We
also now allow links to file names with spaces in them, and disallow
'/' in a filename.