Ls_Var_Log mapper documentation with sample input and usage code
This update changes the Ls_Var_Log mapper to use the FileListing mapper class,
but still provides a few properties that are relied on in some rules and
tests.

This has been tested against the CEEPH repository and all tests pass.  Code
test coverage is at 100%.

This also includes a small edit to the FileListing mapper to keep the file
name in a variable and use that rather than referring into the dictionary each
time it's used in file line parsing.