OSADispatcherLog tests at 100% code coverage
Test last() method on regular and truncated outputs.  Also two minor improvements
to the mapper: a better way to get the groups found in the regular expression
into the line dictionary, and test for parsing success by looking for client_ip key
in the line, rather than testing for its value (which may throw a KeyError).