Uname mapper code coverage at 92% (#181)
* Removing parse_uname function from uname mapper module

The parse_uname function was a left-over legacy of the old mapper function
system, which was included in the util module of Falafel for historic
reasons.  This has been removed in favour of the new class-based mapper
system.

* Changing from falafel.util.uname to falafel.mappers.uname

* Moved tests from util/uname into mapper tests

Tests for:

* pad_release
* fixed_by (including RHEL 5 2.6.32 kernels)
* unknown_release
* from_release
* all equality and inequality operators
* instantiation from a boot2docker kernel

* Uname mapper code coverage at 92%

Fixed many inconsistencies, added lots more tests, simplified some of the
logic, made sure that CEEPH tests all passed.

Remaining section that is not tested is the parse_uname() function provided by
the uname module.  Why does this exist?  What's it used for?  It looks
completely redundant.