Improvements to SshDConfig mapper (#197)
* Improvements to SshDConfig mapper

Adding some more functionality to the SshDConfig mapper to allow us to
reduce the number of local mappers in the security rules.

* Store the whole line as it appeared in the config file for later.
* Provide a ``get_line(keyword, default)`` method to get the last line of
  configuration that defines this keyword.  This returns the raw, unparsed
  line as stored above.  Takes an optional string, defaulting to '', which
  will be used to construct a fake config line with ``# implicit default``
  in it to show the user a line as if it was defined.
* The ``last()`` method now takes a ``default`` keyword argument, defaulting
  to ``None``.  If no configuration lines with this keyword were found, the
  default value is returned instead.

These two combined make it much easier to test for keywords in the
configuration file, find their values, and report the original line (if any)
to the user.

* Show mapper documentation

* Add ``get_values`` method to get all values in various formats

get_values makes it easier for rules to get all the data from a keyword
that allows multiple values per line - e.g. ListenAddress, Cipher,
AllowedEnv and HostKey.

It takes a 'default' option, so if the value is not found this is
used instead.

It takes a 'join_with' option, which if given will join the lines found
together into one line with this string.  If not supplied, the array of
values is returned as is.

It takes a 'split_on' option, which if given will then split the string
on the given string into an array.  If not given, the string as joined
is returned.  'split_on' is ignored if 'join_with' is not supplied.