Pam mapper revisions to improve parsing accuracy (#157)
* pam module tests at 100% code coverage

Considerable simplification of the parsing process using regular expressions to
eliminate all lines that PAM would ignore.  Deduplification of PamConf as a
subclass of PamDConf.  PamDConf only throws one exception - when it's been
subclassed but the file name is left blank and therefore the service cannot be
derived.  We now test this, various other parsing errors, and more.  CEEPH
tests also pass 100%.

* Added documentation improvements