Add LVM shared reducer and fixes for LVM mappers (#126)
* Add LVM shared reducer and fixes for LVM mappers

This commit fixes some problems parsing output of the lvs, vgs,
and pvs commands in columnar format with headings. New mappers were
added to parse the table format `LvsHeadings`, `PvsHeadings`, and
`VgsHeadings`.  A shared reducer `Lvm` was also added to combine all of
the LVM information regardless of the source.  Tests and documentation
were added for all mappers and the reducer.