Alternatives output into parser module 1.x (#255)
* Rename display_java to more general alternatives parser module

* Rename display_java RST file to more general alternatives parser module

* Updated documentation link file

* Use AlternativesOutput base class for JavaAlternatives parser class

There's a lot of information possible in the output of `alternatives --display`,
and this parser class captures it all.  This also allows other parsers, say for
`alternatives --display mta`.  It allows for more analysis of the output to find
problems.

This also deprecates the `default_java` parser function, whose output is now
available in the `link` property of the JavaAlternatives class.

* Removing extra, unused test data

* Simplifying parsing of paths and priorities