Add component metadata and subgraph evaluation.
Component types and declarations can be enriched with arbitrary
metadata by declaring a type like component = new_component_type(type_metadata=)
or directly in the component declaration with @component(metadata=...).
The metadata in the component declaration inherits from the type and overrides
keys in case of collision.

dr now includes get_subgraphs which takes a dependency graph and yields all
of the disjoint subgraphs within it. This functionality is used within
run_incremental to yield brokers for each independent subgraph given a starting
graph. This should help with memory consumption when executing large numbers of
largely independent components.