Add get_dir() method for the mount point associated with a path (#261)
One common test is 'what is the disk space associated with this path?', which
finds the mount point that underlies that path. This routine tries to find
that path, iteratively removing path segments from the end of the given path
to find a mount point.  Absolute paths will always resolve at '/', which will
always be mounted, so a mount point will always be found.  Relative paths will
never resolve to an absolute mount and will return None.

Also tidied the logic in the line parsing, which restores 100% code coverage
in testing.