Add pyOpenSSL as a project dependency
This is a temporary workaround to get the tests passing again.

I'm adding it to the runtime dependencies for now, but once we figure
out if and how we are going to split out the dependencies for the
client, then this dep will go there.

Also, I think we should reconsider making this library a requirement at
all.