Clean up and properly handle startup options
Come up with simple protocol to communicate to parent process (i.e.
client wrapper script) that collection and upload does not need to
happen and instead print the output of the child process.

The basic indicator is a string prefixed with "INIT: " printed to
stdout.

The logging and of methods called in handle_startup() are cleaned up a
bit as well.