Enhance unsplit_lines to allow keeping continuation character for 1.x branch (#348)
* Enhance unsplit_lines to allow keeping continuation character

Some continuation formats use the 'line continuation' character as a
'list item separator' character as well.  For these formats, its
useful to keep the continuation character on the end of the line.
This update adds a `keep_cont_char` option, defaulting to False.  If
set to True, the line continuation character will be left on the end
of the line.

Tests included, code coverage is 100%.

* Adding documentation for keep_cont_char option

* Added examples of cont_char and keep_cont_char option usage)