Added keyword search to PAM configuration parser for master branch (#366)
* Added keyword search to PAM configuration parser

* Add a keyword search method to PAM configuration objects.  This maps each
  objects' properties over to dictionary keys, with the 'control_flags'
  key being a key/value dictionary instead of a list.  This makes it possible
  to search by keyword/value arguments using the regular `keyword_search`
  helper function, but it returns the original PAMConfEntry objects rather
  than the dictionary.
* Improved documentation.
* Full test coverage of new functionality.

* Using relative paths for Insights libraries

* Rewrote comprehensions as loops for Python 2.6

* Fix handling of None in module arguments for search

For historical reasons, if no module arguments are declared then the 'module_args'
property gets the value None.  Because this is non-searchable, we replace this
with '' while building our modified row list.

Also adding a 'module_args_dict' as a dictionary of key-value pairs to save
rule writers having to use optlist_to_dict themselves.  This will be back-ported
to version 1.x