update test-connection option to return cleaner output
Changed some test-connection lines to log as debug rather than info
in order to clean up the output. Mainly removes the printing of the
cert from stdout, but it still logs it. This cleans the output up a
bit for a customer so they can more easily read what's going on.