Ps parsers update for 1.x branch (#744)
* New PsAuxww parser that takes over from the previous parsers

There are four different parsers for ``ps`` output, depending on the flags used
to invoke it.  All of them collect slightly different stuff.  Some are no
longer in use.  Some collect insufficient data (e.g. a truncated command line).
Some contain useless repeated flags ('ww' gets extra wide output, 'www' does
not do anything more).

So of course the solution to this is to create a fifth parser that does
everything that the other four do, and more.  The aim is to replace all use of
the other four parsers in all rule code, and then deprecate the use of the
previous parsers.

The new parser is ``PsAuxww``.  It collects all the statistics that ``PsAux``,
``PsAuxcww`` and ``PsAuxwww`` do, and gets full command lines (unlike ``PsAux``
and ``PsAuxcww``).  It does not get the full start time of the process (which
``PsAxcwwo`` does), but there are no recorded uses of ``PsAxcwwo`` so removing
it just makes the code simpler.

In addition, ``PsAuxww`` implements a ``search`` method that allows keyword-
based searching on any of the columns in the output in combination.  This,
combined with the ``__contains`` keyword search functionality, makes it easy
to perform sophisticated searches on the process list.

* Adding ps_auxww spec for updated PsAuxww parser

* Use new PsAuxww parser

* Use new PsAuxww parser

* Only include complete rows when parsing content, simplify in and fuzzy_match

* Test for reading 'ps auxcww' data

* IpcsSemaphores needs to take both the new and old Ps parsers

Rule tests still pass the data in by spec name, not by path.  So to be backward
compatible with rules, IpcsSemaphores needs to read from either the new PsAuxww
or old PsAuxcww parsers.

* Minor indentation fix for ParseException

* Updating documentation with PsAuxww and warnings on other deprecated parsers

* Returning __contains__ to 'proc in running'

* Use self instead of cls in init functions

* Wrap tests of deprecated parsers in warnings catcher