Add 'deprecated' function to log use of deprecated parsers for master branch (#726)
* The 'deprecated' function and documentation

* Making deprecation warnings fatal and pending deprecations always warn

* Removing deprecated functionality from tests

* DeprecationWarnings are now non-fatal and are the only type of deprecation

This brings our use of `DeprecationWarning` into line with other parts of the
Python ecosystem - they issue a DeprecationWarning if something is to be
removed in a subsequent release, and the feature is removed in that later
release.  The PendingDeprecation is no longer used.  DeprecationWarning is
no longer a fatal error, but will be seen in tests.

* Updating documentation for deprecation process