Refine the HttpdV parser (#1018)
* Refine the HttpdV parser

In the new framework 3.0, HttpdV parser is based on the new datasource
which can get the unique and correct `httpd` command by analyzing the
`ps auxww` output. In RHEL6, the command could be `httpd`, `httpd.worker`
or `httpd.event`, but only one of them will be collected. In RHEL7, only
`httpd` is provided.
- Remove the specs: `httpd_event_V` and `httpd_worker_V`
- Update the `httpd_V` spec to use wildcard in `insights_archive.py`
- Raise `SkipComponent` in the `HttpdWorkerV` and `HttpdEventV` when necessary
- Mark the `HttpdV` combiner as deprecated
- Update the examples in `HttpdV` parser as per doctest

* Fix flake8 errors and a tiny order fix

* Update the httpd_v spec to get the correct httpd command

- ignore multiple httpd instance temporary

* Revert unrelated typo

* Change back to use ps_auxww