New parser for `engine-config --all` command (#1019)
Very early version of parser for command `engine-config`. I intended
to design parser where one could pass the key on-the-fly and the
command will execute it as

```
engine-config --get=KEY
```

to output value. But this seem not possible at the moment as the specs
file run the command and collect the output as an insight archive.

The implementation still works on the output from `engine-config
--all`. The parser tries its best to get 'value' & 'version' for 'key'
specified(Please see examples in doc-string). It ignores keys where is
fails(try...expect). If key with same name exist, it creates a list of
values and versions accordingly.

Includes:
- Tests
- Documentation & Examples

Signed-off-by: Sachin Patil <psachin@redhat.com>