Create regression tests
Created tests suites for the changed pieces of code. The tests are not
DRY and do not test the actual introduced methods. The reason for
that is that they test that the behavior didn’t change with the
refactoring. It is not intended to leave them like this. They are left
in this commit for reference.