Limit maximum pyOpenSSL version
The most recent version of pyOpenSSL fails to install on Python
2.6. Its cryptography dependency cannot be satisfied. Added a limit to
the last version that still supports 2.6. Since the limit is only from
the upper side, it still allows to use pyOpenSSL 0.13.1 installed on
RHEL6 from RPM.

Along with pyOpenSSL, version of pycparser has been limited too. The
most recent version does not support Python 2.6.

This change allows to use virtualenv on RHEL6 in Jenkins the same way it is
used in the other pipeline branches.