--from-file reads uploader.json properly
Currently if we pass the 'from-file' option to insights-client, it's
unable to run.

For example, try running this:

```bash
sudo BYPASS_GPG=True insights-client --no-gpg  --verbose --offline
--from-file compliance.json
```

where compliance.json is:

```json
{
    "files": [
        {
            "file": "/var/lib/insights/latest-compliance-results.xml",
            "pattern": []
        },
        {
            "file": "/var/lib/insights/latest-compliance-results.html",
            "pattern": []
        }
    ],
    "version": "2017-08-31T17:10:37.456370",
    "sig": "asdasd",
    "uploader.json": "/etc/insights-client/.cache.json",
    "file": "/home/dlobatog/workspace/insights/insights-client/compliance.json"
}
```

uploader.json is read as a string instead of reading the file,
therefore it does not read the rules properly. It should be reading
the file instead as shown in this commit