Remove the collection_remove_file constant
The collection_remove_file_constant is no longer needed. It is used
only as a default config value. Its presence might be confusing and
tempt to use it instead of the dynamic value from the config. Removed
and improved the tests so they pass a config value instead of patching
this constant.