Run pytest in virtualenv on RHEL6 (#1451)
A Jenkins pytest task for RHEL 6 did not use a virtual
environment. That was because of broken dependencies in setup.py. With
those fixed, it’s possible to use a virtualenv on RHEL 6 the same way
it is used on RHEL 7.

Temporarily added pyOpenSSL and pycparser to the Jenkins pipeline to
install manually on RHEL 6. This makes the installation work even
though the dependencies are buggy now.

(cherry picked from commit cd1f764c303fd9da89c4816eaef90fa19e301416)