Fix remove file usage in collection tests (#1447)
test_get_rm_conf.test_no_file did not work properly: it tested the
default value of the remote_file constant instead of the patched
one. Fixed this to be consistend with other tests.

test_collect.patch_isfile was patching the isfile method and the
collection_remove_file constant at once, which was not very
transparent and violated SRP. Split into two methods.

Changed mocked removed files in test_collect so their names are not so
confusing. It looked like configuration file names, but they are files
to be removed. Also renamed the variables to make their purpose
clearer.