Add method to get count of processes with given string (#1548)
This commit adds a method to the Ps parser that returns the number
of occurencies of commands with a given string. This is very much needed
and useful when writing rules for OpenStack because we have some performance
considerations wehn using a large number of worker processes for services.
For example, the recommendation is to cap the woker count at
ming(processorcount/2, 12) to avoid issues with database conenctions etc, so
we need the parser to provide this data in order to write rules.

(cherry picked from commit 4fa61b9c6f91e5fa5a08739ea1d5e03e847d60ae)