Add component, type, and key to rule results in evaluators' json response. (#1586)
Each rule response type gets an element in its own section of the doc. This
commit adds the rule's full name (package and function name) as "component",
the type of the response (rule, pass, fingerprint, etc.) as "type", and the
first argument of `make_*` as "key" to each response at the level just above
"details".

The fully qualified name lowers the chance of collision between rule keys and
so is a better unique key in other systems.

This change lifts metadata about the rule and response out of the details
section where it's mixed with pydata.

This change is backward compatible with the previous doc format.