Delay setting permissions when untarring. Adjust how roots are found. (#1665)
Some archives have directories with permissions that make it impossible
to write to them. Without delaying the permission set on those directories,
any files they contain will fail to extract due to PermissiongDenied.

Archives that have multiple directories in the root evaluated the common
prefix to be too high in the directory structure. This change looks for
marker files (sos_commands, insights_commands, etc.) and take the immediate
parent as the archive root.