Allow filtering by full module name and tags.
All components have a tags keyword in their decorators, and the values defined
there can be overridden in configuration. This PR provides a --tags option to
insights-run that allows users to express which rules they want to run based
on how they're tagged. Tag expressions are a small language that allows exact
match, regular expression match, and grouping along with operators for negation,
"and", and "or".

For example, `insights run -p telemetry --tags 'network & apache & !mod_cluster'`
means select rules that have the network and apache tags but not mod_cluster.

insights-run also has a -k option that applies the same expression language to the
package structure of loaded components. If -k and --tags are used at the same
time, the effect is an "and" of their combined logic. In other words, load
components with -p, filter them by package with -k, and select from what's left
with --tags.

For example, `insights run -p telemetry -k '/networking & !/bond'` means load all
of the telemetry rules and run only those that have networking in the full module
name but that don't have bond in the full module name.

The same -k and --tags options are available for insights-info.

Signed-off-by: Christopher Sams <csams@redhat.com>