Execution contexts register themselves and decide if they handle an archive.
A context gets registered automatically when subclassing ExecutionContext.

Each subclass should specify a marker file relative to the archive root
that is unique to the archive type the context handles.

Execution context classes are consulted in reverse of the order in which
they are loaded, so it's possible to load your own context in place of one
provided by core.

```
class MyContext(ExecutionContext):
	marker = "mycontext.txt"
```

Signed-off-by: Christopher Sams <csams@redhat.com>