Encode tags dict as a list of dicts
This encoding process loops over each key/value pair in tags and
performs one of three actions:
1) If the value is a list, the encoding function is called recursively
on each element in the list. The key is repeated for each element.
The results are collected and chained into a list, which is then
returned.
2) If the value is a dict, the encoding function is called recursively
on each key/value pair in the dict. The key is first prefixed with the
"parent" key plus a ":" (colon) as a scope delimiter. The results are
collected and chained into a list, which is then returned.
3) All over values return a dictionary with two entries: "key" has a
value of the current entry key. "value" has a value of the current
entry value. An additional "namespace" entry is added, with a
value pulled from `constants.app_name`.

Signed-off-by: Link Dupont <link@sub-pop.net>