Add package check in CapsuleVersion combiner (#2351)
satellite-capsule package can be found on satellite server also due to
misconfiguration. If present CapsuleVersion combiner will evaluate for
satellite server also. Added a check in combiner to skip component if
foreman package is found on system as foreman package will be only on
satellite and not on capsule server.

Signed-off-by: Rohan Arora <roarora@redhat.com>