Use utf8 in glob (#2372)
* Change encoding in glob_file

If a file contains a unicode symbol, it gets corrupted in rules that try
to read it. By encoding with utf-8 instead of ascii, the symbol is
properly decoded by the rule.

This ties to this JIRA issue:
https://projects.engineering.redhat.com/browse/RHCLOUD-3610

Signed-off-by: Stephen Adams <tsadams@gmail.com>

* Use codecs.open for glob_files to ensure proper encoding

It's possible that some files could contain characters that are
improperly decoded in rules. This establishes utf-8 across the board so
that it's less likely to misinterpret a character.

Signed-off-by: Stephen Adams <tsadams@gmail.com>

* Update _stream function in glob_file

Signed-off-by: Stephen Adams <tsadams@gmail.com>