Exit if the remove.conf file is invalid
We should exit if the remove.conf file is invalid instead of continuing
to process the archive. This always happened if the --validate flag was
applied, but if not, the client would continue to run even with a bad
remove.conf. This exits on a bad remove.conf if the file exists at all.

RHCLOUD-2649

Signed-off-by: Stephen Adams <tsadams@gmail.com>