Delete archive dir (#2463)
* Always delete the archive dir during cleanup

We should get rid of the archive dir during cleanup. We can still keep
the resulting file which can easily be extracted if there is a need to
see what is inside

Signed-off-by: Stephen Adams <tsadams@gmail.com>

* Delete the archive tmp dir after obfuscation

The soscleaner uses a different dir when obfuscation happens. We need to
account for that an delete the empty dir left behind

Signed-off-by: Stephen Adams <tsadams@gmail.com>

* remove archive creation with obfuscate

Since soscleaner takes over the archive creation when obfuscate=True, we
can bypass the archive tmp directory creation.

Signed-off-by: Stephen Adams <tsadams@gmail.com>

* Remove extra archive delete command

We don't need the extra command

Signed-off-by: Stephen Adams <tsadams@gmail.com>