Fix %f regex for get_after
strptime supports Microseconds with %f but not Milliseconds. When logs
have milliseconds, %f still works fine to convert to datetime object as zeros
are padded to right.

When LogFileOutput uses %f in time_format but there are 3 digit milliseconds,
get_after function does not work correctly as it looks for a 6 digit
number. This change adjusts the regex to {1,6} digits for get_after to
work.

Source : Comments for %f in following documentation :
https://docs.python.org/2/library/datetime.html#strftime-and-strptime-behavior

Signed-off-by: Rohan Arora <roarora@redhat.com>