Silence user warning when importing pkg_resources (#2639)
* Silence user warning when importing pkg_resources

When importing this package, sys.path warnings are recorded on RHEL6.
These are noise and can be safely ignored. They are currently causing
cron emails to be sent every time insights runs on a RHEL6 box.

RHCLOUD-7066
BZ 1847104

Signed-off-by: Stephen Adams <tsadams@gmail.com>

* Better fix for UserWarning

pkg_resources was imported above a line that already ignored warnings. I
moved the import of it down into the function that used it to take
advantage of the already implemented warning ignore line.

Signed-off-by: Stephen Adams <tsadams@gmail.com>