Couple PID files write and delete
Originally, the PID files were written on the wrapper init call before
the phases run. Their removal was however scheduled in the collect and
output phase. That resulted in the files not being deleted if this
phase is not run.

Moved the registration of the shutdown handler that deletes the file,
to the init phase. Like this, whenever the file gets created, its
deletion is registered. They should no longer remain undeleted.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>