RHCLOUD-9230 Delete pid files on exit (#2795)
* Couple PID files write and delete

Originally, the PID files were written on the wrapper init call before
the phases run. Their removal was however scheduled in the collect and
output phase. That resulted in the files not being deleted if this
phase is not run.

Moved the registration of the shutdown handler that deletes the file,
to the init phase. Like this, whenever the file gets created, its
deletion is registered. They should no longer remain undeleted.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Fix PID file write tests

The tests verifying that the PID files are written were broken. They
succeeded even if the file is not written. It was because incorrect
unittest.mock usage. Fixed.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Test delete PID files (not) being registered

Added new tests verifying that a function that deleted the PID files
is (not) registered when InsightsClient is initialized, depending on
whether is is called from a phase. Not testing the actual filesystem
delete as the actual write is untested too.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

Co-authored-by: Jeremy Crafts <gravitypriest@users.noreply.github.com>