Use Response in check-in tests
get_canonical_facts does not return a pure dict, but a Response object
instead. This makes an actual difference, because this object has its
own string representation that used to fail before fixing. Replaced
plain dicts in the check-in tests to use this type instead. Like that
the tests pass only if the stringification works, which is what
happens in the executive code.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>