Ultralight checkins (#2807)
* DRY Inventory URL to a property

The Inventory API URL is used a few times in connection.py. DRIED this
by storing it in a property derived from api_url. This is going to be
used by the check-in action too.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* First working check-in prototype

Added the new check-in method all the way down from the CLI argument
to the connection object method. The check-in really happens with the
machine ID, only printing out the result without any error check.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* DRY API request failure

Extracted HTTP request error catching and logging. This will be used
by the new checkin method.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Correctly process responses

Process the check-in responses and report success on success and
failure on failure accordingly. Log some useful information along
the way.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Rename Machine ID to Insights ID

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Test the new check-in method

Created tests for the new InsightsConnection.checkin method. It tests:

- The API request URL, headers and body.
- Fallback from Canonical Facts to Insights ID.
- Raised exceptions and return values.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Add tests for Inventory URL

Created a new test suite for the InsightsConnection constructor.
There, the Inventory URL is composed. Tested that this is done
correctly.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Add InsightsClient.checkin tests

Added simple tests verifying that the Insights Client calls
InsightsConnection.checkin.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Do not check-in in offline mode

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Write tests for Insights Phase

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Remove non-ASCII character

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Lint

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Make check-in work with recent API changes

The Inventory check-in API changed to be more restful.

- POST is used to create a check-in.
- 201 Created is returned on success.
- 404 Not Found is return if host not found.

Changed the check-in code to reflect these changes. 400 Bad Request
and 200 OK are now considered unknown responses.

Also fixed some test comments.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Move the Canonical Facts to the root object

Canonical facts are no longer posted as a separate property. They are
first-class properties of the request.

Patched Response.__str__ method to support results of
get_canonical_facts. It removes the type key from the result and
the string method does not expect that.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Use Response in check-in tests

get_canonical_facts does not return a pure dict, but a Response object
instead. This makes an actual difference, because this object has its
own string representation that used to fail before fixing. Replaced
plain dicts in the check-in tests to use this type instead. Like that
the tests pass only if the stringification works, which is what
happens in the executive code.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

* Test Response stringification

There was a bug in Response.__str__ that caused the canonical facts
stringification to fail with KeyError. Added a test for the fix.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>

Co-authored-by: Jeremy Crafts <gravitypriest@users.noreply.github.com>