Remove debug log messages generated when commands fail during serialization.
Sometimes commands fail during archive creation. core logs the failures at
debug level, but they're generally harmless and causing confusion when people
inspect the collection log. Since we also capture the failures in the metadata
for each component, we can remove the debug statement and not lose the info.

Fixes #2913

Signed-off-by: Christopher Sams <csams@redhat.com>