Remove debug call that prints unhelpful tracebacks during component loading. (#2914)
When dr is asked for a component, it will first ensure that relevant module is
loaded. Since the component can be a module attribute or a class attribute, the
loader tries to get it as each, and this can cause harmless ImportErrors. A
debug statement was added to log them for visibility, but in practice they're
causing confusion in the logs.

Fixes #2912

Signed-off-by: Christopher Sams <csams@redhat.com>
(cherry picked from commit 1750b2df4f0b9913ff1ebf35fa2d7cec64fd9905)