fix: collect version_info in core3 archives (#2918)
* fix: collect version_info in core3 archives

Looks like we need this spec defined for core3 archives. Ran a test
locally and this is the only reason I can spot that we wouldn't be
populating this field in some cases.

Signed-off-by: Stephen Adams <tsadams@gmail.com>

* fix: add version_info to default.py

Fix a bug where core and egg version info are missing from
system_profile. Puptoo needs to use this spec to extract it

Signed-off-by: Stephen Adams <tsadams@gmail.com>