Fix checkin URL with legacy upload
The Inventory URL is composed incorrectly if legacy upload
is enabled. There is a workaround for this in place that
disables legacy upload for actions that use the Inventory
API. Added the check-in to the list of such actions.

Removed show_status from the workaround action list,
because the legacy upload config option does not affect
it in any way.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>