core: introduce commands
Introduce a way to register and execute commands that can filter,
search, compare or otherwise process data stored by a component based on
input data provided by the user.

A component is just a method of another Component registered by
insights. A decorator is introduced to help register these commands and
expose them to the insights shell.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>