Do not compare by is (not) with literals (#3139)
Comparing string literals by is yields a SyntaxWarning on newer
Pythons. This happens in insights.client.connection that uses is/is
not to compare with string and number literals. Not only this issues
warnings, but is itself incorrect behavior that may lead to bugs.
Hence the warnings. Fixed and tested.

Signed-off-by: Štěpán Tomsa <stomsa@redhat.com>
(cherry picked from commit e6b31604b641cc082c0bb4558e152ce0ea3aedad)