fix: use LC_ALL=C.UTF-8 for subscription-manager
The default environment for "simple_command" includes LC_ALL=C, which
means ASCII (and not unicode); in case the language of the system is set
to a non-English locale that uses unicode characters (e.g. non-Latin1
languages), then subscription-manager will fail to output that with
encoding issues (e.g. [1]).

As a simple solution, enforce an unicode English locale for
subscription-manager, which should avoid those encoding issues.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2074745

Signed-off-by: Pino Toscano <ptoscano@redhat.com>