feat: Allow listdir to return relative paths
The `listdir` spec factory returned only basenames of listed files and
directories. However, the path of files and directories may contain
information that is not available anywhere else and that rules may need
to access.

This commit adds a `path_format` argument to the `listdir` spec factory
that allows the spec factory to generate a data source that returns
paths relative to the execution context root.

Signed-off-by: Jan Holeček <71874510+jholecek-rh@users.noreply.github.com>