feat: Split listdir to listdir and listglob spec factories (#3694)
* fix: Add unit tests for listdir spec factory

The tests declare the intended behavior of the spec factory that future
extensions and/or refactoring should not change inadvertently.

Signed-off-by: Jan Holeček <71874510+jholecek-rh@users.noreply.github.com>

* feat: Simplify behavior of listdir spec factory

The exact behavior of specs created using the listdir spec factory
was difficult to predict because it depended on a number of subtle
details:

- ContentException was raised instead of an empty list in some cases, an
  empty list was returned in other cases.
- The ignore filter was applied to returned values in some cases and
  to absolute paths in other cases.
- The specs returned only basenames of files and directories even for
  a complex glob (e.g. 'some/*/file') where returning relative paths
  makes more sense.

This commit simplifies the spec factory by
- Removing support for globs (no known specs rely on this
  functionality)
- Raising ContentException only if the specificed path is not an existing
  directory; an empty list will be returned for empty directories
  (SkipComponent will be raised one step later for known listdir specs;
  the change should not affect any rules).

Signed-off-by: Jan Holeček <71874510+jholecek-rh@users.noreply.github.com>

* feat: Add listglob spec factory

Glob patterns were originally supported by the listdir spec factory,
causing its behavior difficult to predict without knowing the internals.

The previous commit removed support for globs from the listdir spec
factory to simplify its behavior.

A dedicated listglob spec factory has a more predictable behavior while
keeping the implementation simple. In most cases, users will know if
they want to list a directory or a glob pattern. An amphibious spec
factory is not necessary.

Signed-off-by: Jan Holeček <71874510+jholecek-rh@users.noreply.github.com>

---------

Signed-off-by: Jan Holeček <71874510+jholecek-rh@users.noreply.github.com>